+++ 6lbr/ws/ws_pan_info_storage.h
@@ -21,7 +21,7 @@
 
 #include <stdint.h>
 
-void ws_pan_info_storage_read(uint16_t *bsi, int *pan_id, uint16_t *pan_version, uint16_t *lfn_version, char network_name[33]);
+void ws_pan_info_storage_read(int *bsi, int *pan_id, uint16_t *pan_version, uint16_t *lfn_version, char network_name[33]);
 void ws_pan_info_storage_write(uint16_t bsi, uint16_t pan_id, uint16_t pan_version, uint16_t lfn_version, const char network_name[33]);
 
 #endif
